refactor ChartPopover
