+++ packages/charts/react-charts/library/src/components/ChartHoverCard/index.ts
@@ -1,0 +1,2 @@
+export * from './ChartHoverCard';
+export * from './ChartHoverCard.types';
