+++ packages/charts/react-charts/library/src/components/ChartHoverCard/useChartHoverCardStyles.styles.ts
@@ -1,0 +1,169 @@
+import { makeStyles, mergeClasses, shorthands } from '@griffel/react';
+import { HighContrastSelectorBlack, HighContrastSelector } from '../../utilities/index';
+import { SlotClassNames } from '@fluentui/react-utilities/src/index';
+import { tokens, typographyStyles } from '@fluentui/react-theme';
+import { ChartHoverCardProps, ChartHoverCardStyles } from './ChartHoverCard.types';
+
+/**
+ * @internal
+ */
+export const hoverCardClassNames: SlotClassNames<ChartHoverCardStyles> = {
+  calloutContentRoot: 'fui-cart__calloutContentRoot',
+  calloutDateTimeContainer: 'fui-cart__calloutDateTimeContainer',
+  calloutContentX: 'fui-cart__calloutContentX',
+  calloutBlockContainer: 'fui-cart__calloutBlockContainer',
+  calloutBlockContainertoDrawShapefalse: 'fui-cart__calloutBlockContainertoDrawShapefalse',
+  calloutBlockContainertoDrawShapetrue: 'fui-cart__calloutBlockContainertoDrawShapetrue',
+  shapeStyles: 'fui-cart__shapeStyles',
+  calloutlegendText: 'fui-cart__calloutlegendText',
+  calloutContentY: 'fui-cart__calloutContentY',
+  descriptionMessage: 'fui-cart__descriptionMessage',
+  ratio: 'fui-cart__ratio',
+  numerator: 'fui-cart__numerator',
+  denominator: 'fui-cart__denominator',
+  calloutInfoContainer: 'fui-cart__calloutInfoContainer',
+};
+
+/**
+ * Base Styles
+ */
+const useStyles = makeStyles({
+  calloutContentRoot: {
+    display: 'contents',
+    overflow: 'hidden',
+    ...shorthands.padding('11px 16px 10px 16px'),
+    backgroundColor: tokens.colorNeutralBackground1,
+    backgroundBlendMode: 'normal, luminosity',
+  },
+  calloutDateTimeContainer: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  calloutContentX: {
+    ...typographyStyles.caption1,
+    opacity: '0.8',
+    color: tokens.colorNeutralForeground2,
+  },
+  calloutBlockContainer: {
+    color: tokens.colorNeutralForeground2,
+  },
+  calloutBlockContainerCartesian: {
+    ...typographyStyles.caption1,
+    marginTop: '13px',
+  },
+  calloutBlockContainerNonCartesian: {
+    fontSize: tokens.fontSizeHero700,
+    lineHeight: '22px',
+    '& selectors': {
+      [HighContrastSelector]: {
+        forcedColorAdjust: 'none',
+      },
+    },
+  },
+  calloutBlockContainertoDrawShapefalse: {
+    '& selectors': {
+      [HighContrastSelector]: {
+        forcedColorAdjust: 'none',
+      },
+    },
+    paddingLeft: tokens.spacingHorizontalS,
+  },
+  calloutBlockContainertoDrawShapetrue: { display: 'inline-grid' },
+  shapeStyles: {
+    marginRight: tokens.spacingHorizontalS,
+  },
+  calloutLegendText: {
+    ...typographyStyles.caption1,
+    color: tokens.colorNeutralForeground2,
+    '& selectors': {
+      [HighContrastSelectorBlack]: {
+        color: 'rgb(255, 255, 255)',
+      },
+    },
+  },
+  calloutContentY: {
+    '& selectors': {
+      [HighContrastSelectorBlack]: {
+        color: 'rgb(255, 255, 255)',
+      },
+    },
+  },
+  calloutContentYCartesian: {
+    ...typographyStyles.subtitle2Stronger,
+  },
+  calloutContentYNonCartesian: {
+    ...typographyStyles.title2,
+  },
+  descriptionMessage: {
+    ...typographyStyles.caption1,
+    color: tokens.colorNeutralForeground2,
+    marginTop: tokens.spacingVerticalMNudge,
+    paddingTop: tokens.spacingVerticalMNudge,
+    ...shorthands.borderTop(`1px solid ${tokens.colorNeutralStroke2}`),
+  },
+  ratio: {
+    ...typographyStyles.caption2,
+    marginLeft: tokens.spacingHorizontalSNudge,
+    color: tokens.colorNeutralForeground1,
+  },
+  numerator: {
+    ...typographyStyles.caption2Strong,
+  },
+  denominator: {
+    ...typographyStyles.caption2Strong,
+  },
+  calloutInfoContainer: {
+    paddingLeft: tokens.spacingHorizontalS,
+  },
+});
+
+export const useChartHoverCardStyles_unstable = (props: ChartHoverCardProps): ChartHoverCardStyles => {
+  const { isCartesian } = props;
+  const baseStyles = useStyles();
+  return {
+    calloutContentRoot: mergeClasses(
+      hoverCardClassNames.calloutContentRoot,
+      baseStyles.calloutContentRoot /*props.styles?. calloutContentRoot*/,
+    ),
+    calloutDateTimeContainer: mergeClasses(
+      hoverCardClassNames.calloutDateTimeContainer,
+      baseStyles.calloutDateTimeContainer /*props.styles?.calloutDateTimeContainer*/,
+    ),
+    calloutContentX: mergeClasses(
+      hoverCardClassNames.calloutContentX,
+      baseStyles.calloutContentX /*props.styles?.calloutContentX*/,
+    ),
+    calloutBlockContainer: mergeClasses(
+      hoverCardClassNames.calloutBlockContainer,
+      baseStyles.calloutBlockContainer /*props.styles?.calloutBlockContainerCartesian*/,
+      isCartesian ? baseStyles.calloutBlockContainerCartesian : baseStyles.calloutBlockContainerNonCartesian,
+    ),
+    calloutBlockContainertoDrawShapefalse: mergeClasses(
+      hoverCardClassNames.calloutBlockContainertoDrawShapefalse,
+      baseStyles.calloutBlockContainertoDrawShapefalse /*props.styles?.calloutBlockContainertoDrawShapefalse*/,
+    ),
+    calloutBlockContainertoDrawShapetrue: mergeClasses(
+      hoverCardClassNames.calloutBlockContainertoDrawShapetrue,
+      baseStyles.calloutBlockContainertoDrawShapetrue /*props.styles?.calloutBlockContainertoDrawShapetrue*/,
+    ),
+    shapeStyles: mergeClasses(hoverCardClassNames.shapeStyles, baseStyles.shapeStyles /*props.styles?.shapeStyles*/),
+    calloutlegendText: mergeClasses(
+      hoverCardClassNames.calloutlegendText,
+      baseStyles.calloutLegendText /*props.styles?.calloutlegendText*/,
+    ),
+    calloutContentY: mergeClasses(
+      hoverCardClassNames.calloutContentY,
+      baseStyles.calloutContentY /*props.styles?.calloutContentYNonCartesian*/,
+      isCartesian ? baseStyles.calloutContentYCartesian : baseStyles.calloutContentYNonCartesian,
+    ),
+    descriptionMessage: mergeClasses(
+      hoverCardClassNames.descriptionMessage,
+      baseStyles.descriptionMessage /*props.styles?. descriptionMessage*/,
+    ),
+    ratio: mergeClasses(hoverCardClassNames.ratio, baseStyles.ratio /*props.styles?.ratio*/),
+    numerator: mergeClasses(hoverCardClassNames.numerator, baseStyles.numerator /*props.styles?.numerator*/),
+    denominator: mergeClasses(hoverCardClassNames.denominator, baseStyles.denominator /*props.styles?.denominator*/),
+    calloutInfoContainer: mergeClasses(hoverCardClassNames.calloutInfoContainer, baseStyles.calloutInfoContainer),
+  };
+};
