+++ packages/charts/react-charts/library/src/components/CommonComponents/CartesianChart.types.ts
@@ -431,12 +431,6 @@ export interface CartesianChartProps {
    * @default false
    */
   hideTickOverlap?: boolean;
-
-  /**
-   * Define a custom callout props override
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  customProps?: (dataPointCalloutProps: any) => ChartPopoverProps;
 }
 
 export interface YValueHover {
