+++ packages/charts/react-charts/library/src/components/VerticalBarChart/VerticalBarChart.tsx
@@ -953,10 +953,7 @@ export const VerticalBarChart: React.FunctionComponent<VerticalBarChartProps> = 
     isCalloutForStack: _isHavingLine && _noLegendHighlighted(),
     culture: props.culture ?? 'en-us',
     isCartesian: true,
-    customCallout: {
-      customizedCallout: _getCustomizedCallout() != null ? _getCustomizedCallout()! : undefined,
-      customCalloutProps: props.customProps ? props.customProps(dataPointCalloutProps!) : undefined,
-    },
+    customizedCallout: _getCustomizedCallout() != null ? _getCustomizedCallout()! : undefined,
   };
 
   const tickParams = {
