+++ packages/charts/react-charts/library/src/index.ts
@@ -8,3 +8,4 @@ export * from './types/index';
 export * from './Sparkline';
 export * from './utilities/colors';
 export * from './Popover';
+export * from './ChartHoverCard';
